refactor(event): remove duplicated module prefix in event file names
