+++ pumpkin/src/entity/item.rs
@@ -1,5 +1,5 @@
 use crate::{
-    entity::EntityBaseFuture, plugin::player::player_pickup_arrow::PlayerPickupArrowEvent,
+    entity::EntityBaseFuture, plugin::player::pickup_arrow::PlayerPickupArrowEvent,
     server::Server,
 };
 use core::f32;
