+++ pumpkin/src/plugin/api/events/player/pickup_arrow.rs
