+++ pumpkin/src/plugin/api/events/player/recipe_discover.rs
