+++ pumpkin/src/plugin/api/events/player/spawn_location.rs
